+++ sechecker/modules/find_node_types.c
@@ -294,7 +294,7 @@ int find_node_types_run(sechk_module_t * mod, apol_policy_t * policy, void *arg 
 
 	for (i = 0; i < apol_vector_get_size(nodecon_vector); i++) {
 		char *type_name;
-		int j;
+		size_t j;
 		qpol_context_t *context;
 		qpol_type_t *context_type;
 		qpol_nodecon_t *nodecon = apol_vector_get_element(nodecon_vector, i);
@@ -373,7 +373,7 @@ int find_node_types_print(sechk_module_t * mod, apol_policy_t * policy, void *ar
 	unsigned char outformat = 0x00;
 	sechk_item_t *item = NULL;
 	sechk_proof_t *proof = NULL;
-	int i = 0, j = 0, k = 0, num_items = 0;
+	size_t i = 0, j = 0, k = 0, num_items = 0;
 	qpol_type_t *type;
 	qpol_policy_t *q = apol_policy_get_qpol(policy);
 	char *type_name;
